backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arrow==1.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
exceptiongroup==1.3.1
fastapi==0.135.1
flexcache==0.3
flexparser==0.4
h11==0.16.0
idna==3.11
joblib==1.5.3
numpy==2.2.6
pandas==2.3.3
Pint==0.24.4
platformdirs==4.9.4
pydantic==2.12.5
pydantic_core==2.41.5
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
starlette==0.52.1
stravalib==2.4
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
polyline==2.0.4
python-multipart>=0.0.9
